Give connector_builder its own main.py
